# colexecerror: avoid debug.Stack in CatchVectorizedRuntimeError