--- a/pkg/sql/sem/builtins/fixed_oids.go
+++ b/pkg/sql/sem/builtins/fixed_oids.go
@@ -2573,6 +2573,7 @@ var builtinOidsArray = []string{
 	2605: `merge_aggregated_stmt_metadata(arg1: jsonb) -> jsonb`,
 	2606: `crdb_internal.protect_mvcc_history(timestamp: decimal, expiration_window: interval, description: string) -> int`,
 	2607: `crdb_internal.extend_mvcc_history_protection(job_id: int) -> void`,
+	2608: `crdb_internal.force_panic(msg: string, mode: string) -> int`,
 }
 
 var builtinOidsBySignature map[string]oid.Oid