@michae2 2024-04-30
Benchmark results are here: https://gist.github.com/michae2/4406203dbafc5749ad6a02f8b0ec268e