@mgartner 2024-04-30
*[`pkg/sql/colexecerror/error.go` line 48 at r2](https://reviewable.io/reviews/cockroachdb/cockroach/123277#-Nwjsmd39mfZH5SQIhI8:-NwlGkBdCNwO1qa7LWSb:b3tjets) ([raw file](https://github.com/cockroachdb/cockroach/blob/2917eecda0b954cef30c803348db8d97656430bb/pkg/sql/colexecerror/error.go#L48)):*
<details><summary><i>Previously, rafiss (Rafi Shamim) wrote…</i></summary><blockquote>

`errors.Is[Any]` requires the error (or any error in the cause chain) to exactly equal a reference error.

`errors.As` checks if the error (or any error in the cause chain) is assignable to the value pointed at by the target. 

in this case, since there is no "singleton" `notInternalError`, we need to use `As`.
</blockquote></details>

Thanks for the explanation Rafi!


<!-- Sent from Reviewable.io -->