@michae2 2024-05-01
I've added benchmark results to the PR description, collected using:

```
N=1 BENCHTIMEOUT=24h PKG=./pkg/sql/colexecerror BENCHES=CatchVectorizedRuntimeError ./scripts/bench catch~3 catch
```

(I had to run with N=1 to avoid exhausting ports on my gceworker, so the benchstat output is missing the nice comparison, but the absolute numbers are accurate.)