@yuzefovich 2024-05-01
For posterity, here are the number I got using `scripts/bench` with the old `benchstat` on the gceworker (comparing 1st commit against 4th):
```
name                                                             old time/op    new time/op    delta
CatchVectorizedRuntimeError/noError                                4.75ns ± 0%    4.75ns ± 0%     ~     (p=0.752 n=10+10)
CatchVectorizedRuntimeError/expected                               18.9µs ± 1%     0.5µs ± 0%  -97.47%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/storage                                16.1µs ± 2%     0.4µs ± 0%  -97.71%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/contextCanceled                        16.2µs ± 1%     0.6µs ± 0%  -96.17%  (p=0.000 n=10+9)
CatchVectorizedRuntimeError/internalWithCode                       18.9µs ± 1%     4.4µs ± 1%  -76.55%  (p=0.000 n=9+10)
CatchVectorizedRuntimeError/internal                               23.1µs ± 1%     8.8µs ± 1%  -61.77%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/runtime                                24.3µs ± 1%    10.4µs ± 1%  -57.29%  (p=0.000 n=10+9)
SQLCatchVectorizedRuntimeError/conns=1/noError                      466µs ± 4%     466µs ± 6%     ~     (p=0.739 n=10+10)
SQLCatchVectorizedRuntimeError/conns=1/expectedWithCode             967µs ± 2%     516µs ± 3%  -46.70%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=1/expectedAssertion           1.37ms ± 3%    0.89ms ± 2%  -34.78%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=1/internalAssertion           1.21ms ± 4%    0.79ms ± 3%  -35.12%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=1/internalIndexOutOfRange     1.27ms ± 3%    0.85ms ± 4%  -33.45%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=1/internalDivideByZero        1.24ms ± 2%    0.83ms ± 2%  -33.64%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=1/contextCanceled              905µs ± 7%     484µs ± 4%  -46.53%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=20/noError                     457µs ± 5%     466µs ± 6%     ~     (p=0.278 n=10+9)
SQLCatchVectorizedRuntimeError/conns=20/expectedWithCode            989µs ± 8%     546µs ± 4%  -44.82%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=20/expectedAssertion          1.41ms ± 5%    0.91ms ±13%  -35.48%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=20/internalAssertion          1.34ms ± 4%    0.82ms ± 4%  -38.89%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=20/internalIndexOutOfRange    1.31ms ±12%    0.94ms ± 3%  -27.96%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=20/internalDivideByZero       1.13ms ± 1%    0.94ms ± 3%  -17.17%  (p=0.000 n=9+10)
SQLCatchVectorizedRuntimeError/conns=20/contextCanceled             878µs ± 2%     456µs ± 5%  -48.04%  (p=0.000 n=9+10)
SQLCatchVectorizedRuntimeError/conns=50/noError                     443µs ± 6%     434µs ± 5%     ~     (p=0.247 n=10+10)
SQLCatchVectorizedRuntimeError/conns=50/expectedWithCode            932µs ± 2%     486µs ± 1%  -47.92%  (p=0.000 n=10+9)
SQLCatchVectorizedRuntimeError/conns=50/expectedAssertion          1.25ms ± 2%    0.77ms ± 5%  -38.32%  (p=0.000 n=9+10)
SQLCatchVectorizedRuntimeError/conns=50/internalAssertion          1.17ms ± 6%    0.70ms ± 4%  -40.14%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=50/internalIndexOutOfRange    1.24ms ± 8%    0.76ms ± 2%  -38.61%  (p=0.000 n=10+8)
SQLCatchVectorizedRuntimeError/conns=50/internalDivideByZero       1.21ms ± 4%    0.77ms ± 2%  -36.86%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=50/contextCanceled             943µs ± 7%     485µs ± 6%  -48.51%  (p=0.000 n=9+10)

name                                                             old alloc/op   new alloc/op   delta
CatchVectorizedRuntimeError/noError                                 0.00B          0.00B          ~     (all equal)
CatchVectorizedRuntimeError/expected                               9.83kB ± 0%    0.03kB ± 0%  -99.67%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/storage                                8.88kB ± 0%    0.03kB ± 0%  -99.64%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/contextCanceled                        8.84kB ± 0%    0.04kB ± 0%  -99.55%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/internalWithCode                       9.82kB ± 0%    1.54kB ± 0%  -84.27%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/internal                               10.7kB ± 0%     3.0kB ± 0%  -71.81%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/runtime                                11.2kB ± 0%     2.7kB ± 0%  -75.77%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=1/noError                     84.3kB ± 4%    85.1kB ± 5%     ~     (p=0.549 n=9+10)
SQLCatchVectorizedRuntimeError/conns=1/expectedWithCode             137kB ± 1%      87kB ± 1%  -35.93%  (p=0.000 n=8+8)
SQLCatchVectorizedRuntimeError/conns=1/expectedAssertion            374kB ± 4%     311kB ± 0%  -17.05%  (p=0.000 n=10+9)
SQLCatchVectorizedRuntimeError/conns=1/internalAssertion            297kB ± 0%     253kB ± 4%  -14.77%  (p=0.000 n=8+9)
SQLCatchVectorizedRuntimeError/conns=1/internalIndexOutOfRange      319kB ± 0%     272kB ± 0%  -14.75%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=1/internalDivideByZero         314kB ± 0%     267kB ± 0%  -15.06%  (p=0.000 n=8+8)
SQLCatchVectorizedRuntimeError/conns=1/contextCanceled              130kB ± 0%      83kB ± 0%  -36.57%  (p=0.000 n=8+9)
SQLCatchVectorizedRuntimeError/conns=20/noError                    83.2kB ± 0%    85.9kB ±10%     ~     (p=0.737 n=9+10)
SQLCatchVectorizedRuntimeError/conns=20/expectedWithCode            137kB ± 1%      87kB ± 0%  -36.29%  (p=0.000 n=9+8)
SQLCatchVectorizedRuntimeError/conns=20/expectedAssertion           376kB ± 4%     310kB ± 0%  -17.40%  (p=0.000 n=10+8)
SQLCatchVectorizedRuntimeError/conns=20/internalAssertion           297kB ± 0%     251kB ± 0%  -15.59%  (p=0.000 n=8+10)
SQLCatchVectorizedRuntimeError/conns=20/internalIndexOutOfRange     319kB ± 0%     272kB ± 0%  -14.77%  (p=0.000 n=10+8)
SQLCatchVectorizedRuntimeError/conns=20/internalDivideByZero        314kB ± 0%     267kB ± 0%  -15.04%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=20/contextCanceled             131kB ± 0%      83kB ± 0%  -36.87%  (p=0.000 n=8+8)
SQLCatchVectorizedRuntimeError/conns=50/noError                    87.0kB ± 8%    83.4kB ± 0%     ~     (p=0.128 n=10+8)
SQLCatchVectorizedRuntimeError/conns=50/expectedWithCode            137kB ± 1%      87kB ± 0%  -36.32%  (p=0.000 n=10+9)
SQLCatchVectorizedRuntimeError/conns=50/expectedAssertion           370kB ± 0%     311kB ± 0%  -16.11%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=50/internalAssertion           307kB ± 9%     257kB ± 6%  -16.09%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=50/internalIndexOutOfRange     325kB ± 4%     272kB ± 0%  -16.17%  (p=0.000 n=9+8)
SQLCatchVectorizedRuntimeError/conns=50/internalDivideByZero        316kB ± 0%     268kB ± 0%  -15.24%  (p=0.000 n=8+9)
SQLCatchVectorizedRuntimeError/conns=50/contextCanceled             134kB ±11%      83kB ± 1%  -38.13%  (p=0.000 n=9+9)

name                                                             old allocs/op  new allocs/op  delta
CatchVectorizedRuntimeError/noError                                  0.00           0.00          ~     (all equal)
CatchVectorizedRuntimeError/expected                                 36.0 ± 0%       3.0 ± 0%  -91.67%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/storage                                  11.0 ± 0%       3.0 ± 0%  -72.73%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/contextCanceled                          11.0 ± 0%       4.0 ± 0%  -63.64%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/internalWithCode                         35.0 ± 0%      37.0 ± 0%   +5.71%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/internal                                 46.0 ± 0%      59.0 ± 0%  +28.26%  (p=0.000 n=10+10)
CatchVectorizedRuntimeError/runtime                                  58.0 ± 0%      51.0 ± 0%  -12.07%  (p=0.000 n=10+10)
SQLCatchVectorizedRuntimeError/conns=1/noError                        839 ± 1%       850 ± 4%     ~     (p=0.362 n=8+10)
SQLCatchVectorizedRuntimeError/conns=1/expectedWithCode             1.00k ± 0%     0.95k ± 0%   -4.46%  (p=0.000 n=8+8)
SQLCatchVectorizedRuntimeError/conns=1/expectedAssertion            1.75k ± 6%     1.57k ± 0%  -10.37%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=1/internalAssertion            1.19k ± 0%     1.25k ± 7%   +5.16%  (p=0.000 n=8+9)
SQLCatchVectorizedRuntimeError/conns=1/internalIndexOutOfRange      1.41k ± 0%     1.41k ± 0%     ~     (p=0.912 n=9+9)
SQLCatchVectorizedRuntimeError/conns=1/internalDivideByZero         1.41k ± 1%     1.41k ± 0%     ~     (p=0.625 n=8+8)
SQLCatchVectorizedRuntimeError/conns=1/contextCanceled                873 ± 0%       864 ± 0%   -0.99%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=20/noError                       837 ± 0%       854 ± 6%     ~     (p=0.288 n=9+10)
SQLCatchVectorizedRuntimeError/conns=20/expectedWithCode            1.00k ± 0%     0.95k ± 0%   -4.87%  (p=0.000 n=9+8)
SQLCatchVectorizedRuntimeError/conns=20/expectedAssertion           1.79k ± 7%     1.58k ± 0%  -11.80%  (p=0.000 n=10+8)
SQLCatchVectorizedRuntimeError/conns=20/internalAssertion           1.19k ± 0%     1.23k ± 0%   +3.45%  (p=0.000 n=8+10)
SQLCatchVectorizedRuntimeError/conns=20/internalIndexOutOfRange     1.43k ± 0%     1.42k ± 0%   -0.21%  (p=0.010 n=10+8)
SQLCatchVectorizedRuntimeError/conns=20/internalDivideByZero        1.42k ± 0%     1.42k ± 0%     ~     (p=0.601 n=9+9)
SQLCatchVectorizedRuntimeError/conns=20/contextCanceled               885 ± 0%       868 ± 0%   -1.87%  (p=0.000 n=9+8)
SQLCatchVectorizedRuntimeError/conns=50/noError                       870 ± 6%       844 ± 0%     ~     (p=0.145 n=10+8)
SQLCatchVectorizedRuntimeError/conns=50/expectedWithCode            1.01k ± 1%     0.96k ± 0%   -5.55%  (p=0.000 n=10+9)
SQLCatchVectorizedRuntimeError/conns=50/expectedAssertion           1.76k ± 0%     1.59k ± 0%   -9.82%  (p=0.000 n=9+9)
SQLCatchVectorizedRuntimeError/conns=50/internalAssertion           1.28k ±10%     1.28k ± 7%     ~     (p=0.342 n=10+10)
SQLCatchVectorizedRuntimeError/conns=50/internalIndexOutOfRange     1.52k ±12%     1.44k ± 0%   -5.54%  (p=0.000 n=10+8)
SQLCatchVectorizedRuntimeError/conns=50/internalDivideByZero        1.45k ± 1%     1.44k ± 1%   -1.07%  (p=0.000 n=8+9)
SQLCatchVectorizedRuntimeError/conns=50/contextCanceled               944 ±17%       884 ± 1%   -6.37%  (p=0.000 n=9+9)
```